+++ core/src/test/java/uk/gov/justice/services/core/it/SenderRequesterHandlerIT.java
@@ -49,8 +49,8 @@ import uk.gov.justice.services.core.json.PayloadExtractor;
 import uk.gov.justice.services.core.json.SchemaCatalogAwareJsonSchemaValidator;
 import uk.gov.justice.services.core.mapping.ActionNameToMediaTypesMappingObserver;
 import uk.gov.justice.services.core.mapping.DefaultMediaTypesMappingCache;
+import uk.gov.justice.services.core.mapping.DefaultNameToMediaTypeConverter;
 import uk.gov.justice.services.core.mapping.DefaultSchemaIdMappingCache;
-import uk.gov.justice.services.core.mapping.NameToMediaTypeConverter;
 import uk.gov.justice.services.core.mapping.SchemaIdMappingObserver;
 import uk.gov.justice.services.core.requester.Requester;
 import uk.gov.justice.services.core.requester.RequesterProducer;
@@ -151,7 +151,7 @@ public class SenderRequesterHandlerIT {
 
             SchemaCatalogAwareJsonSchemaValidator.class,
             PayloadExtractor.class,
-            NameToMediaTypeConverter.class,
+            DefaultNameToMediaTypeConverter.class,
             DefaultSchemaIdMappingCache.class,
             SchemaIdMappingObserver.class,