+++ generators/direct-adapter-core/src/test/java/uk/gov/justice/services/adapter/direct/DirectAdapterIT.java
@@ -44,8 +44,8 @@ import uk.gov.justice.services.core.json.PayloadExtractor;
 import uk.gov.justice.services.core.json.SchemaCatalogAwareJsonSchemaValidator;
 import uk.gov.justice.services.core.mapping.ActionNameToMediaTypesMappingObserver;
 import uk.gov.justice.services.core.mapping.DefaultMediaTypesMappingCache;
+import uk.gov.justice.services.core.mapping.DefaultNameToMediaTypeConverter;
 import uk.gov.justice.services.core.mapping.DefaultSchemaIdMappingCache;
-import uk.gov.justice.services.core.mapping.NameToMediaTypeConverter;
 import uk.gov.justice.services.core.mapping.SchemaIdMappingObserver;
 import uk.gov.justice.services.core.requester.RequesterProducer;
 import uk.gov.justice.services.core.sender.SenderProducer;
@@ -138,7 +138,7 @@ public class DirectAdapterIT {
             UrlConverter.class,
             SchemaCatalogAwareJsonSchemaValidator.class,
             PayloadExtractor.class,
-            NameToMediaTypeConverter.class,
+            DefaultNameToMediaTypeConverter.class,
             DefaultSchemaIdMappingCache.class,
             SchemaIdMappingObserver.class,
             ClasspathResourceLoader.class,