+++ generators/rest-adapter-core/src/main/java/uk/gov/justice/services/adapter/rest/mapper/BadRequestExceptionMapper.java
@@ -3,9 +3,10 @@ package uk.gov.justice.services.adapter.rest.mapper;
 import static javax.json.Json.createObjectBuilder;
 import static javax.ws.rs.core.Response.Status.BAD_REQUEST;
 import static javax.ws.rs.core.Response.status;
-import static uk.gov.justice.services.core.json.JsonValidationLogger.toJsonObject;
 
 import uk.gov.justice.services.adapter.rest.exception.BadRequestException;
+import uk.gov.justice.services.core.json.JsonSchemaValidationException;
+import uk.gov.justice.services.core.json.JsonValidationLoggerHelper;
 
 import javax.inject.Inject;
 import javax.json.JsonObjectBuilder;
@@ -13,7 +14,6 @@ import javax.ws.rs.core.Response;
 import javax.ws.rs.ext.ExceptionMapper;
 import javax.ws.rs.ext.Provider;
 
-import org.everit.json.schema.ValidationException;
 import org.slf4j.Logger;
 
 @Provider
@@ -22,14 +22,15 @@ public class BadRequestExceptionMapper implements ExceptionMapper<BadRequestExce
     @Inject
     Logger logger;
 
+    @Inject
+    JsonValidationLoggerHelper jsonValidationLoggerHelper;
+
     @Override
     public Response toResponse(final BadRequestException exception) {
         logger.debug("Bad Request", exception);
-
         final JsonObjectBuilder builder = createObjectBuilder().add("error", exception.getMessage());
-
-        if (exception.getCause() instanceof ValidationException) {
-            builder.add("validationErrors", toJsonObject((ValidationException) exception.getCause()));
+        if (exception.getCause() instanceof JsonSchemaValidationException) {
+            builder.add("validationErrors", jsonValidationLoggerHelper.toJsonObject((JsonSchemaValidationException) exception.getCause()));
         }
 
         return status(BAD_REQUEST)