+++ CHANGELOG.md
@@ -6,6 +6,10 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 ## [Unreleased]
 
 ### Changed
+ - Move LoggerUtils to utilities as part of rest-adapter dependency update
+ - Remove dependencies on core in messaging-adapter-core and partly in rest-adapter-core
+ - Move DefaultJsonValidationLoggerHelper to core
+ - Use JsonSchemaValidatonException instead of org.everit.json.schema.ValidationException
  - SPI for InterceptorContext.interceptorContextWithInput()
 
 ###